Chaser chases player but its ugly AF

+++ Assets/_Scripts/Core/CoreComponent/Movement.cs
@@ -76,7 +76,7 @@
         
         #endregion
         
-        public void CheckIfShouldFlip(int xInput)
+        public void CheckIfShouldFlip(float xInput)
         {
             if(xInput != 0 && xInput != FacingDirection)
             {
